Add files via upload
This was a class project I worked on. The instructions are as follows:

ABC Bank of Nigeria just contacted you to create an ATM automation solution to allow users (bank 
customers) to perform basic financial transactions. The bank will integrate the software with the 
ATM’s hardware at a later time. So instead of writing your program to run on the ATM, you are 
expected to write it to run on your desktop or mobile phone using your preferred Python IDE. This 
version is expected to simulate a real-life ATM. The functionalities/requirements of the ATM solution 
you are expected to come up with is as follows:

(1) A welcome message
(2) A prompt to enter account number and PIN nos. 
(3) Verification of account number and PIN nos. The account number is 10 digits and PIN nos 5 
digits. If both entries are incorrect, user should be prompted to re-enter. After 3 trials and 
verification is not confirmed, the user should be logged off/ban from continuing/proceeding 
further. 
(4) If the user enters a valid account and pin numbers, the main menu screen should be 
displayed.
(5) If user chooses option 1 to view balance, the screen must display the balance of the user. To 
do so, the balance must be retrieved from a variable that stores the user’s balance. 
(6) The operation of option 2 (withdraw cash) is as follows:
a. A display screen that shows standard withdrawal amount is displayed (see Figure 2). 
The user has to choose one of the options, the 6th one terminates the withdrawal 
process/transaction.
b. If the amount chosen is greater than the user’s account balance, the user should be 
informed and prompted to enter a smaller amount using the withdrawal menu 
screen. Otherwise, the amount chosen should be deducted from the user’s account
balance. Then the user is notified of his/her balance and prompted if it wants to 
withdrawal again.
(7) The following steps describe the actions that occur when the user enters option 3 of the 
main menu(Fig. 1) to make a deposit:
a. The screen prompts the user to enter a deposit amount or type 0 (zero) to cancel.
b. The user enters a deposit amount or 0 using the keypad. 
c. If the user specifies a deposit amount, the ATM proceeds to Step d. If the user 
chooses to cancel the transaction (by entering 0), the deposit transaction would be 
terminated. 
d. The ATM credits the deposit amount to the user’s account in the balance variable.
(i.e., adds the deposit amount to the user’s account balance).
(8) After the system successfully executes a transaction, it should give an appropriate message. 
If the user exits the system, the screen should display a thank you message, then display the 
welcome message for the next user or to restart the application for another 
transaction/functionality.
For simulation/testing purpose.
Create different customer(s) storing their names, account number, pin and balance (money). 
E.g. (customer 1’s detail)
Customer1_name = "Thomas shelby" 
Customer1_acct_n0 = "0123456789" 
Customer1_Pin = "00001" 
Customer1_acct_bal = 50000